requirements.txt
# Python dependencies for mock data generation
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
